backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Microsoft Graph API
httpx==0.25.1

# ML/AI (Free models) - Optional for now
# sentence-transformers==2.2.2  # Uncomment when implementing near-duplicates
# scikit-learn==1.3.2
# numpy==1.24.3

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-pptx==0.6.23

# Text similarity
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

